--- a/.gitmodules
+++ b/.gitmodules
@@ -85,7 +85,7 @@
 [submodule "third_party/ideep"]
     ignore = dirty
     path = third_party/ideep
-    url = https://github.com/intel/ideep
+    url = https://github.com/zhuhaozhe/ideep
 [submodule "third_party/nccl/nccl"]
     ignore = dirty
     path = third_party/nccl/nccl