--- a/aten/src/ATen/native/native_functions.yaml
+++ b/aten/src/ATen/native/native_functions.yaml
@@ -2420,18 +2420,22 @@
 - func: mkldnn_linear(Tensor self, Tensor weight, Tensor? bias=None) -> Tensor
   python_module: nn
   dispatch:
+    CPU: mkldnn_linear
     MkldnnCPU: mkldnn_linear
 
 - func: mkldnn_linear_backward_input(int[] input_size, Tensor grad_output, Tensor weight) -> Tensor
   dispatch:
+    CPU: mkldnn_linear_backward_input
     MkldnnCPU: mkldnn_linear_backward_input
 
 - func: mkldnn_linear_backward_weights(Tensor grad_output, Tensor input, Tensor weight, bool bias_defined) -> (Tensor, Tensor)
   dispatch:
+    CPU: mkldnn_linear_backward_weights
     MkldnnCPU: mkldnn_linear_backward_weights
 
 - func: mkldnn_linear_backward(Tensor self, Tensor grad_output, Tensor weight, bool[3] output_mask) -> (Tensor, Tensor, Tensor)
   dispatch:
+    CPU: mkldnn_linear_backward
     MkldnnCPU: mkldnn_linear_backward
 
 - func: fbgemm_linear_int8_weight_fp32_activation(Tensor input, Tensor weight, Tensor packed, Tensor col_offsets, Scalar weight_scale, Scalar weight_zero_point, Tensor bias) -> Tensor
@@ -5166,6 +5170,12 @@
   python_module: nn
   dispatch:
     MkldnnCPU: mkldnn_reorder_conv3d_weight
+
+- func: mkldnn_reorder_linear_weight(Tensor self) -> Tensor
+  variants: function
+  python_module: nn
+  dispatch:
+    MkldnnCPU: mkldnn_reorder_linear_weight
 
 - func: to_mkldnn_backward(Tensor grad, Tensor input) -> Tensor