--- a/test/test_mkldnn.py
+++ b/test/test_mkldnn.py
@@ -940,17 +940,21 @@ class TestMkldnn(TestCase):
 
         for bias in [True, False]:
             linear = torch.nn.Linear(in_features, out_features, bias=bias).float()
+            cpu_linear_bf16 = copy.deepcopy(linear).bfloat16()
             mkldnn_linear = mkldnn_utils.to_mkldnn(copy.deepcopy(linear))
-            mkldnn_linear_bf16 = mkldnn_utils.to_mkldnn(copy.deepcopy(linear), torch.bfloat16)
             if has_bf16_support():
                 y = mkldnn_linear(x.to_mkldnn()).to_dense()
-                y_bf16 = mkldnn_linear_bf16(x_bf16.to_mkldnn()).to_dense(torch.float32)
-                self.assertEqual(y, y_bf16, atol=1e-1, rtol=1e-3)
+                mkldnn_linear_bf16 = mkldnn_utils.to_mkldnn(copy.deepcopy(linear), torch.bfloat16)
+                y_bf16 = mkldnn_linear_bf16(x_bf16.to_mkldnn()).to_dense()
+                cpu_y_bf16 = cpu_linear_bf16(x_bf16)
+
+                self.assertEqual(y, y_bf16.float(), atol=1e-1, rtol=1e-3)
+                self.assertEqual(cpu_y_bf16, y_bf16, atol=1e-1, rtol=1e-3)
             else:
-                msg = "mkldnn_linear: bf16 path needs the cpu support avx512bw, avx512vl and avx512dq"
+                msg = "mkldnn_reorder_linear_weight: bf16 path needs the cpu support avx512bw, avx512vl and avx512dq"
                 self.assertRaisesRegex(RuntimeError,
                                        msg,
-                                       lambda: mkldnn_linear_bf16(x_bf16.to_mkldnn()))
+                                       lambda: mkldnn_utils.to_mkldnn(copy.deepcopy(linear), torch.bfloat16))
 
     def test_softmax(self):
         x = torch.randn(3, 4, 5, dtype=torch.float32) * 10