## Goal: Integrate mkldnn bf16 Gemm to pytorch

### Changes for Mkldnn-linear (with mkldnn inner product pritimitve):
https://oneapi-src.github.io/oneDNN/group__dnnl__api__inner__product.html
- Input tensor type:
    - Before: input must be mkldnn tensor
    - Now: input can be a bf16 cpu tensor (Do not convert bf16 cpu tensor to bf16 mkldnn tensor can save a mem copy)
- Weight packs:
    - Before: to_mkldnn(model) only convert nn.Linear to nn.mkldnn_linear (weight will not be packed)
    - Now: weight will be packed after converted to mkldnn-linear to avoid run-time format conversion on each iter

### BF16 Inference:
### User Case1:
```
import torch
LL = torch.nn.Linear(256, 256).bfloat16()
# forward
LL(torch.rand(16, 256).bfloat16())
```
### User Case2 (preferred):
```
import torch
LL = torch.nn.Linear(256, 256).bfloat16()
mkldnn_LL = mkldnn_utils.to_mkldnn(LL, dtype=torch.bfloat16) # will pack weight from plain format to expected block format to accelerate gemm 
# forward
mkldnn_LL(torch.rand(16, 256).bfloat16())
```

### BF16 training:
For training, we suppose the batch size is big, so the reorder of weight for each iter will not be a big overhead.
```
import torch
LL = torch.nn.Linear(256, 256).bfloat16()
# forward
y = LL(torch.rand(16, 256).bfloat16())
# backward
y.mean().backward()
```

## Performace gain (fw)

Note: data is collected on 28core/per socket CPX with using pytorch operator benchmarks: https://github.com/pytorch/pytorch/tree/master/benchmarks/operator_benchmark (with adding bfloat16 dtype)

### use 1 thread on 1 core
### torch.linear (N, IN) * (IN, OUT) + (OUT, ) 
| impl |32x32x32| 64x64x64 | 128x128x128| 256x256x256| 512x512x512|1024x1024x1024|
|:---:| :---: | :---: | :---: | :---: | :---: | :---: |
| aten-fp32| 12.379us|17.067us| 38.726us|219.284us|1.456ms|11.310ms|
| aten-bf16 |  166.797us|1.278ms|10.061ms|80.748ms|0.646s|too-long-to-run-over|
| mkldnn-bf16 |  28.658us| 31.817us|46.882|140.936us|0.821ms|6.398ms|
| mkldnn-bf16-packed |  28.165us|30.669us|41.678us|117.074us|0.698ms|6.006ms|


### use 1 socket (28 thread, 28 core)
| impl | 256x256x256| 512x512x512|1024x1024x1024| 2048x2048x2048|4096x4096x4096|
|:---:| :---: | :---: | :---: | :---: | :---: |
| aten-fp32| 45.413us|118.624us|758.04us|7.468ms|43.143ms|
| mkldnn-bf16 | 57.145us|122.302us|459.665us|4.237ms|270.66ms|
| mkldnn-bf16-packed | 43.621us|75.252us|369.91us|2.636ms|250.11ms|

## Performace gain (bw)
### use 1 socket (28 thread, 28 core)
| impl | 4096x256x256| 4096x512x512|4096x1024x1024| 4096x2048x2048|4096x4096x4096|
|:---:| :---: | :---: | :---: | :---: | :---: |
| aten-fp32| 0|0|0|0|0|
| mkldnn-bf16 | 0|0|0|0|0|
| mkldnn-bf16-packed | 0|0|0|0|0|